requirements.txt
beautifulsoup4==4.11.1
openpyxl==3.0.9
PyYAML==6.0
requests_html==0.10.0
